==> notebooks/03_clean.ipynb <==
# Clean data

import os
import numpy as np
import pandas as pd
from datetime import date
import pyarrow as pa
import pyarrow.parquet as pq

# load processed data
df = pq.read_table('../data/processed.parquet').to_pandas()

### Exclude entries with shared name & gender and agegroups > 12 years apart

# Parkrun have been organized in Australia for 12 years and counting so a name/gender cannot be associated with more than 12 years across age groups
df['EXCLUDE_AGE'] = df.groupby(['name','gender'])['agegroup'].transform(lambda x: (int(x.max()[:2]) - int(x.min()[-2:])) > 12)

### Exclude parkruns with less than 100 runs

# Could make more strict to active parkruns only (merge with new data)
df['EXCLUDE_RUN'] = df.groupby('parkrun')['n'].transform(lambda x: x.sum()) < 100

### Other exclusions?
- i.e. more refined age exclusion with for instance one or more agegroups missing between the large gaps -> two people

## save cleaned data

df = df.query('not (EXCLUDE_AGE or EXCLUDE_RUN)').drop(['EXCLUDE_AGE','EXCLUDE_RUN'], axis = 1)
df.info()

table = pa.Table.from_pandas(df, preserve_index = False)
pq.write_table(table, '../data/cleaned.parquet')
